requirements.txt
kafka-python==2.0.2
pandas==2.2.2
plotly==5.24.1
psycopg2-binary==2.9.9
pytest==8.3.3
streamlit==1.39.0
